requirements.txt
fastapi
numpy
Pillow
uvicorn
python-multipart
tensorflow-cpu